!!![BUGFIX] Properly replace uploaded files
In case a file reference is not annotated as
ObjectStorage and a reference should be replaced with a newly
uploaded file, a new file reference was created.

This happened because a new file reference object was
created which Extbase persisted as new file reference row.

The configuration option to remove previously existing
file reference rows was a workaround for that which
had unwanted side effects (see #3).

This change properly handles this situation, by fetching
already existing file references with the persistence manager.

By doing so, the correct file reference is updated instead.

This is breaking your controller code, if you used the remove configuration
option, as the constant is also removed.

Fixes #3